requirements.txt
# Judol Hunter - Web Application Dependencies

# Core Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
aiosqlite>=0.19.0
aiomysql>=0.2.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# Configuration & Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Templates
jinja2>=3.1.2

# Rate Limiting
slowapi>=0.1.9

# HTTP Client & Web Scraping (from original CLI)
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Terminal Output (from original CLI)
rich>=13.0.0

# ASGI Server (production)
gunicorn>=21.0.0
